refactor: porter runner command
