+++ src/runner/porter-command/porterCommandInterface.ts
@@ -1,0 +1,5 @@
+import { EntityManager } from 'typeorm';
+
+export interface PorterCommand {
+  execute(manager: EntityManager): Promise<void>;
+}
